Modify terminal warning message regarding cpu and fast path code
It seems this terminal warning message should be triggered by my
patched code, but it isn't. But in case it ever is, I added a
comment to clarify that the relevant code that would be speeded up
in default babl has been removed from my patched babl,
so there's nothing to speed up.